(Most of) Problem 3 using the backtracking linesearch algorithm
The parameters for this are very fiddly and it doesn't always produce
a great result. It is necessary to reduce the precision, else the
backtracking algorithm struggles to converge in a reasonable amount of
time. That said, it does seem to work well with the parameters I've chosen.

Left to implement: multiple runs across the specified range to try and
find the global minimum in that range.