Fix linesearch algorithm
The choice of p = -∇f/|∇f| does not make sense if the α produced by
the linesearch algorithm is to be used in gradient descent. This is
why the algorithm was not working properly except for values of c
close to 1. Correcting this allows small values of c (e.g. c = 1e-3)
to be used without issue, and allows both the linesearch algorithm and
gradient descent algorithm to converge very quickly.

This is all explained in a large comment block.